@@ -19,11 +19,12 @@
         "php": "^7.2.5",
         "psr/container": "^1.0",
         "symfony/deprecation-contracts": "^2.1",
+        "symfony/polyfill-php80": "^1.15",
         "symfony/service-contracts": "^1.1.6|^2"
     },
     "require-dev": {
         "symfony/yaml": "^4.4|^5.0",
-        "symfony/config": "^5.0",
+        "symfony/config": "^5.1",
         "symfony/expression-language": "^4.4|^5.0"
     },
     "suggest": {
@@ -34,7 +35,7 @@
         "symfony/proxy-manager-bridge": "Generate service proxies to lazy load them"
     },
     "conflict": {
-        "symfony/config": "<5.0",
+        "symfony/config": "<5.1",
         "symfony/finder": "<4.4",
         "symfony/proxy-manager-bridge": "<4.4",
         "symfony/yaml": "<4.4"
